requirements.txt
git+git://github.com/niwibe/django-sse
https://www.djangoproject.com/download/1.7c1/tarball/
Jinja2==2.7.2
MarkupSafe==0.21
Werkzeug==0.9.4
django-bower==5.0.1
https://bitbucket.org/nicfit/eyed3/get/tip.zip#egg=eyeD3-0.7.5
gevent==1.0.1
greenlet==0.4.2
gunicorn==18.0
itsdangerous==0.24
oauthlib==0.6.3
python-magic==0.4.6
python-openid==2.2.5
python-social-auth==0.1.26
redis==2.10.1
requests==2.2.1
requests-oauthlib==0.4.1
six==1.7.2
sse==1.2
wsgiref==0.1.2
django-filer==0.9.10
